refactor: migrate elasticsearch storage implementation to new api
Since "cqp-api" release "0.0.16" it includes all types of storage API.
It make unnecessary dependency on "cqp-core" which is not related to "storage" only to task execution in Akka cluster.
