+++ cqp-storage-elasticsearch/src/main/java/com/quantori/cqp/storage/elasticsearch/ElasticsearchIterator.java
@@ -12,8 +12,8 @@ import co.elastic.clients.elasticsearch.core.SearchRequest;
 import co.elastic.clients.elasticsearch.core.search.Hit;
 import co.elastic.clients.elasticsearch.core.search.ResponseBody;
 import co.elastic.clients.transport.ElasticsearchTransport;
+import com.quantori.cqp.api.SearchIterator;
 import com.quantori.cqp.api.model.Flattened;
-import com.quantori.cqp.core.model.SearchIterator;
 import com.quantori.cqp.storage.elasticsearch.model.MoleculeDocument;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
