+++ cqp-storage-elasticsearch/src/main/java/com/quantori/cqp/storage/elasticsearch/ElasticsearchStorageMolecules.java
@@ -15,24 +15,24 @@ import co.elastic.clients.elasticsearch.core.SearchRequest;
 import co.elastic.clients.elasticsearch.core.search.Hit;
 import co.elastic.clients.json.JsonData;
 import co.elastic.clients.transport.ElasticsearchTransport;
+import com.quantori.cqp.api.Criteria;
+import com.quantori.cqp.api.ItemWriter;
 import com.quantori.cqp.api.MoleculesFingerprintCalculator;
 import com.quantori.cqp.api.MoleculesMatcher;
+import com.quantori.cqp.api.SearchIterator;
 import com.quantori.cqp.api.StorageMolecules;
 import com.quantori.cqp.api.model.ConjunctionCriteria;
+import com.quantori.cqp.api.model.ExactParams;
 import com.quantori.cqp.api.model.FieldCriteria;
 import com.quantori.cqp.api.model.Flattened;
 import com.quantori.cqp.api.model.Library;
 import com.quantori.cqp.api.model.Property;
+import com.quantori.cqp.api.model.SearchProperty;
+import com.quantori.cqp.api.model.SimilarityParams;
+import com.quantori.cqp.api.model.SortParams;
+import com.quantori.cqp.api.model.SubstructureParams;
 import com.quantori.cqp.api.model.upload.Molecule;
 import com.quantori.cqp.api.util.FingerPrintUtilities;
-import com.quantori.cqp.core.model.Criteria;
-import com.quantori.cqp.core.model.ExactParams;
-import com.quantori.cqp.core.model.ItemWriter;
-import com.quantori.cqp.core.model.SearchIterator;
-import com.quantori.cqp.core.model.SearchProperty;
-import com.quantori.cqp.core.model.SimilarityParams;
-import com.quantori.cqp.core.model.SortParams;
-import com.quantori.cqp.core.model.SubstructureParams;
 import com.quantori.cqp.storage.elasticsearch.model.MoleculeDocument;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
