+++ cqp-storage-elasticsearch/src/main/java/com/quantori/cqp/storage/elasticsearch/ElasticsearchStorageReactions.java
@@ -7,18 +7,18 @@ import co.elastic.clients.elasticsearch.core.SearchRequest;
 import co.elastic.clients.elasticsearch.core.SearchResponse;
 import co.elastic.clients.elasticsearch.core.search.Hit;
 import co.elastic.clients.transport.ElasticsearchTransport;
+import com.quantori.cqp.api.ItemWriter;
 import com.quantori.cqp.api.ReactionsFingerprintCalculator;
 import com.quantori.cqp.api.ReactionsMatcher;
+import com.quantori.cqp.api.SearchIterator;
 import com.quantori.cqp.api.StorageReactions;
+import com.quantori.cqp.api.model.ExactParams;
 import com.quantori.cqp.api.model.Flattened;
 import com.quantori.cqp.api.model.Library;
+import com.quantori.cqp.api.model.ReactionParticipantRole;
+import com.quantori.cqp.api.model.SubstructureParams;
 import com.quantori.cqp.api.model.upload.ReactionUploadDocument;
 import com.quantori.cqp.api.util.FingerPrintUtilities;
-import com.quantori.cqp.core.model.ExactParams;
-import com.quantori.cqp.core.model.ItemWriter;
-import com.quantori.cqp.core.model.ReactionParticipantRole;
-import com.quantori.cqp.core.model.SearchIterator;
-import com.quantori.cqp.core.model.SubstructureParams;
 import com.quantori.cqp.storage.elasticsearch.model.ReactionDocument;
 import com.quantori.cqp.storage.elasticsearch.model.ReactionParticipantDocument;
 import lombok.extern.slf4j.Slf4j;
